+++ test/e2e/main.spec.js
@@ -74,7 +74,7 @@ describe('Research Data Finder', function () {
     const maxPatientCountInput = $('#maxPatientCount');
     const loadObservationsBtn = $('#ObservationTabPage-1-loadBtn');
 
-     browser.wait(EC.visibilityOf(maxPatientCountInput));
+    browser.wait(EC.visibilityOf(maxPatientCountInput));
 
     // Random maximum number of Patients from 50 to 100
     maxPatientCountInput.sendKeys(Key.chord(Key.CONTROL, 'a'), 50 + Math.floor(Math.random()*50));
@@ -123,15 +123,16 @@ describe('Research Data Finder', function () {
 
       safeClick(getVisibleButtonByText('Download (in CSV format)'));
 
+      let csvData = null;
       browser
         .wait(function () {
           // Wait until the file has been downloaded.
-          return fs.existsSync(filename);
+          return fs.existsSync(filename) &&
+            (csvData = fs.readFileSync(filename, { encoding: 'utf-8' })).length;
         }, 30000)
         .then(function () {
           // Checks CSV file structure: the file has columns and all lines have the same number of cells.
-          const cellsInRowCount = fs
-              .readFileSync(filename, { encoding: 'utf8' })
+          const cellsInRowCount = csvData
               .replace(/""/g, '')
               .replace(/"[^"]*"/g, '')
               .split('\n')
@@ -176,16 +177,16 @@ describe('Research Data Finder', function () {
 
       safeClick($('#saveCohort'));
 
+      var cohortData = null;
       browser
         .wait(function () {
           // Wait until the file has been downloaded.
-          return fs.existsSync(filename);
+          return fs.existsSync(filename) &&
+           (cohortData = fs.readFileSync(filename, { encoding: 'utf-8' })).length;
         }, 30000)
         .then(function () {
           // Checks JSON file structure.
-          const cohortData = JSON.parse(
-            fs.readFileSync(filename, { encoding: 'utf-8' })
-          );
+          cohortData = JSON.parse(cohortData);
           expect(cohortData && cohortData.data && cohortData.data.length).toBe(
             patientsCount
           );
@@ -251,6 +252,10 @@ describe('Research Data Finder', function () {
           safeClick(element(by.cssContainingText('.tab-link', resourceType)));
         }
       });
+  }
+
+  function waitForAutocompleteDropDown() {
+    browser.wait(EC.presenceOf($('#searchResults[aria-hidden="false"]')));
   }
 
   describe('without criteria(initial state)', function () {
@@ -324,6 +329,8 @@ describe('Research Data Finder', function () {
   });
 
   describe('when adding criteria to Observation resource', function () {
+// Commented out tests are not relevant when using lastn lookup:
+/*
     it('should provide the ability to select a test name and test value', function () {
       const searchParamId = getNextSearchParamId();
       const addCriterionBtn = $('#SearchParameters-1_add_button');
@@ -337,6 +344,7 @@ describe('Research Data Finder', function () {
       resourceInput.sendKeys(Key.chord(Key.CONTROL, 'a'), 'Observation');
       resourceInput.sendKeys(Key.ENTER);
       testNameInput.sendKeys('body height measured');
+      waitForAutocompleteDropDown();
       testNameInput.sendKeys(Key.ARROW_DOWN);
       testNameInput.sendKeys(Key.ENTER);
       browser.wait(EC.presenceOf(testRealValueInput));
@@ -356,6 +364,27 @@ describe('Research Data Finder', function () {
       resourceInput.sendKeys(Key.chord(Key.CONTROL, 'a'), 'Observation');
       resourceInput.sendKeys(Key.ENTER);
       testNameInput.sendKeys('3137-7');
+      testNameInput.sendKeys(Key.ENTER);
+      browser.wait(EC.presenceOf(testRealValueInput));
+      testRealValueInput.sendKeys('63');
+    });
+*/
+
+    it('should provide the ability to select a test name from current FHIR server', function () {
+      const searchParamId = getNextSearchParamId();
+      const addCriterionBtn = $('#SearchParameters-1_add_button');
+      const resourceInput = $(`#${searchParamId}_resource`);
+      const testNameInput = $(`#${searchParamId}-test-name`);
+      const testRealValueInput = $(`#${searchParamId}-test-value`);
+
+      browser.wait(EC.elementToBeClickable(addCriterionBtn));
+      safeClick(addCriterionBtn);
+
+      resourceInput.sendKeys(Key.chord(Key.CONTROL, 'a'), 'Observation');
+      resourceInput.sendKeys(Key.ENTER);
+      testNameInput.sendKeys('height cm');
+      waitForAutocompleteDropDown();
+      testNameInput.sendKeys(Key.ARROW_DOWN);
       testNameInput.sendKeys(Key.ENTER);
       browser.wait(EC.presenceOf(testRealValueInput));
       testRealValueInput.sendKeys('63');